geom_serialAxesGlyph (deprecated)

use ggmulti::geom_serialaxes_glyph, rather than geom_serialAxesGlyph.

Usage

geom_serialAxesGlyph(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  ...,
  serialAxesData,
  sequence = NULL,
  linewidth = 1,
  scaling = c("variable", "data", "observation", "none"),
  axesLayout = c("parallel", "radial"),
  showAxes = FALSE,
  showArea = FALSE,
  showEnclosing = FALSE,
  axesColor = "black",
  bboxColor = "black",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)

Arguments

mapping

Set of aesthetic mappings created by aes(). If specified and inherit.aes = TRUE (the default), it is combined with the default mapping at the top level of the plot. You must supply mapping if there is no plot mapping.

data

The data to be displayed in this layer. There are three options:

If NULL, the default, the data is inherited from the plot data as specified in the call to ggplot().

A data.frame, or other object, will override the plot data. All objects will be fortified to produce a data frame. See fortify() for which variables will be created.

A function will be called with a single argument, the plot data. The return value must be a data.frame, and will be used as the layer data. A function can be created from a formula (e.g. ~ head(.x, 10)).

stat

The statistical transformation to use on the data for this layer, either as a ggproto Geom subclass or as a string naming the stat stripped of the stat_ prefix (e.g. "count" rather than "stat_count")

position

Position adjustment, either as a string naming the adjustment (e.g. "jitter" to use position_jitter), or the result of a call to a position adjustment function. Use the latter if you need to change the settings of the adjustment.

...

Other arguments passed on to layer(). These are often aesthetics, used to set an aesthetic to a fixed value, like colour = "red" or size = 3. They may also be parameters to the paired geom/stat.

serialAxesData

a serial axes numerical data set. If not provided, `geom_point()` will be called.

sequence

A vector with variable names that defines the axes sequence

linewidth

line width of the "glyph" object

scaling

one of data, variable, observation or none (not suggested the layout is the same with data) to specify how the data is scaled.

axesLayout

either "radial" or "parallel"

showAxes

boolean to indicate whether axes should be shown or not

showArea

show area; deprecated now, please set `fill` or `colour` to control the shown area.

showEnclosing

boolean to indicate whether enclosing should be shown or not

axesColor

axes color

bboxColor

bounding box color

na.rm

If FALSE, the default, missing values are removed with a warning. If TRUE, missing values are silently removed.

show.legend

logical. Should this layer be included in the legends? NA, the default, includes if any aesthetics are mapped. FALSE never includes, and TRUE always includes. It can also be a named logical vector to finely select the aesthetics to display.

inherit.aes

If FALSE, overrides the default aesthetics, rather than combining with them. This is most useful for helper functions that define both data and aesthetics and shouldn't inherit behaviour from the default plot specification, e.g. borders().