ggplot serialaxes (deprecated)

For ggSerialAxes, use ggmulti::coord_serialaxes.

Usage

ggSerialAxes(
  ggObj,
  data = NULL,
  axesLabels = NULL,
  showAxes = TRUE,
  showAxesLabels = TRUE,
  scaling = c("variable", "observation", "data", "none"),
  layout = c("parallel", "radial"),
  displayOrder = NULL,
  title = "",
  showLabels = TRUE,
  color = NULL,
  size = NULL,
  showGuides = TRUE,
  showArea = FALSE
)

Arguments

ggObj

A `ggplot` object

data

A data frame for serialaxes. If `NULL`, data must be set in `ggObj`

axesLabels

A vector with variable names that defines the axes sequence.

showAxes

Logical value to indicate whether axes should be shown or not

showAxesLabels

Logical value to indicate whether axes labels should be shown or not

scaling

one of 'variable', 'data', 'observation' or 'none' to specify how the data is scaled. See Details for more information

layout

either "radial" or "parallel"

displayOrder

The display order of the observations.

title

title of the display

showLabels

Logical value to indicate whether label (mainly **title**) should be shown or not

color

Line color

size

Line width

showGuides

Logical value to indicate whether guides should be shown or not

showArea

Logical value to indicate whether to display lines or area

Value

a ggplot object

Examples

if (FALSE) { # \dontrun{
# Blank plot
p <- ggplot(data = mtcars, mapping = aes(colour = factor(cyl)))
# Add serial axes (returns a ggplot object)
g <- ggSerialAxes(p)
g
# An eulerian path of iris variables
# ordSeq <- PairViz::eulerian(4)
ordSeq <- c(1, 2, 3, 1, 4, 2, 3, 4)
ggSerialAxes(
       ggObj = ggplot(data = iris, mapping = aes(colour = Species)),
       axesLabels = colnames(iris)[ordSeq],
       layout = "radial"
)
} # }