Deprecated functions in package loon.ggplot.

The functions listed below are deprecated and will be defunct in the near future. When possible, alternative functions with similar functionality are also mentioned. Help pages for deprecated functions are available at help("geom_imageGlyph-deprecated"), help("geom_polygonGlyph-deprecated"), help("geom_pointrangeGlyph-deprecated"), help("geom_textGlyph-deprecated"), help("geom_serialAxesGlyph-deprecated"), help("lggplot-deprecated"), help("ggSerialAxes-deprecated")