Box scaling in 3D rotation

the variable is scaled to have equal ranges and, when center = TRUE, to be centred by the average of the min and max.

scaleBox(center = TRUE)

Arguments

center
either a logical value or numeric-alike vector of length equal to the number of columns of x, where 'numeric-alike' means that as.numeric(.) will be applied successfully if is.numeric(.) is not true.

Value

A trans object

See also