Create a loon plot from a ggplot2 object

Interactive loon plots from ggplots

loon.ggplot(ggplotObject, activeGeomLayers = integer(0), parent = NULL,
  ggGuides = FALSE, pack = TRUE, tkLabels = NULL,
  exteriorLabelProportion = 1/5, canvasHeight = 700,
  canvasWidth = 850, ...)

Arguments

ggplotObject
a ggplot object

activeGeomLayers
to determine which geom layer is active. Only geom_point() and geom_histogram() can be set as active geom layer(s). (Notice, more than one geom_point() layers can be set as active layers, but only one geom_histogram() can be set as an active geom layer)

parent
parent widget path (Tk toplevel)

ggGuides
logical (default FALSE) to determine whether to draw a ggplot background or not.

pack
logical (default TRUE) to pack widgets. If FALSE, widgets will be produced but won't be packed and so will not appear in the display.

tkLabels
logical (or NULL) to indicate whether the plot(s) are to be wrapped with exterior labels (title, subtitle, xlabel or ylabel) using tk.grid(). If NULL (default), then exterior labels appear only for multiple facets. If TRUE exterior labels appear regardless; if FALSE no exterior labels appear.

exteriorLabelProportion
space assigned to the vertical height/horizontal width of each exterior label expressed as a proportion of a single plot's height/width. Default is 0.2. This is translated to a row/column span = 1 / exteriorLabelProportion for the plot size in tkgrid().

canvasHeight
the height of canvas

canvasWidth
the width of canvas

...
named arguments to modify loon plot states

Examples

 p <- ggplot(mtcars, aes(wt, mpg)) + geom_point()
 g <- loon.ggplot(p)

 # show ggGuides
 p <- ggplot(mpg, aes(class, hwy)) + geom_boxplot()
 g <- loon.ggplot(p, ggGuides = TRUE)

 # tkLabels
p <- ggplot(mtcars) + geom_point(aes(x = wt, y = mpg,
   colour = factor(gear))) + facet_wrap(~am)
g1 <- loon.ggplot(p)

#> default sync is 'pull'

g2 <- loon.ggplot(p, tkLabels = FALSE)
# NOT RUN {
df <- data.frame(
  x = rnorm(120, c(0, 2, 4)),
  y = rnorm(120, c(1, 2, 1)),
  z = letters[1:3]
)
df2 <- dplyr::select(df, -z)
scatterplots <- ggplot(df, aes(x, y)) +
  geom_point(data = df2, colour = "grey70") +
  geom_point(aes(colour = z)) +
  facet_wrap(~z)

# We can select the first geom_point layer to be
# the active layer as in
suppressWarnings(
  lp_scatterplots_active1 <- loon.ggplot(scatterplots,
                               activeGeomLayers = 1,
                               linkingGroup = "test")
)
# Here the grey points are linked (not the coloured ones)

# We can select the second geom_point layer to be
# the active layer as in
lp_scatterplots_active2 <- loon.ggplot(scatterplots, activeGeomLayers = 2)
# Here the colour points are linked

# We can also select the both geom_point layers to be
# the active layer as in
suppressWarnings(
 lp_scatterplots_active12 <- loon.ggplot(scatterplots, activeGeomLayers = c(1,2))
)
# Here the colour points and grey points are both linked
# }