Turn a loon widget to a ggplot object

Create a ggplot object from a loon widget

# S3 method for l_compound
loon2ggplot(
  target,
  asAes = TRUE,
  selectedOnTop = TRUE,
  showNearestColor = FALSE,
  ...
)

# S3 method for l_facet_grid
loon2ggplot(
  target,
  asAes = TRUE,
  selectedOnTop = TRUE,
  showNearestColor = FALSE,
  ...
)

# S3 method for l_facet_wrap
loon2ggplot(
  target,
  asAes = TRUE,
  selectedOnTop = TRUE,
  showNearestColor = FALSE,
  ...
)

# S3 method for l_layer_graph
loon2ggplot(
  target,
  asAes = TRUE,
  selectedOnTop = TRUE,
  showNearestColor = FALSE,
  ...
)

# S3 method for l_layer_histogram
loon2ggplot(
  target,
  asAes = TRUE,
  selectedOnTop = TRUE,
  showNearestColor = FALSE,
  ...
)

# S3 method for l_layer_scatterplot
loon2ggplot(
  target,
  asAes = TRUE,
  selectedOnTop = TRUE,
  showNearestColor = FALSE,
  ...
)

# S3 method for l_pairs
loon2ggplot(
  target,
  asAes = TRUE,
  selectedOnTop = TRUE,
  showNearestColor = FALSE,
  ...
)

# S3 method for l_serialaxes
loon2ggplot(
  target,
  asAes = TRUE,
  selectedOnTop = TRUE,
  showNearestColor = FALSE,
  ...
)

# S3 method for zenLoon
loon2ggplot(
  target,
  asAes = TRUE,
  selectedOnTop = TRUE,
  showNearestColor = FALSE,
  ...
)

loon2ggplot(
  target,
  asAes = TRUE,
  selectedOnTop = TRUE,
  showNearestColor = FALSE,
  ...
)

# S3 method for default
loon2ggplot(
  target,
  asAes = TRUE,
  selectedOnTop = TRUE,
  showNearestColor = FALSE,
  ...
)

# S3 method for l_plot
loon2ggplot(
  target,
  asAes = TRUE,
  selectedOnTop = TRUE,
  showNearestColor = FALSE,
  ...
)

# S3 method for l_hist
loon2ggplot(
  target,
  asAes = TRUE,
  selectedOnTop = TRUE,
  showNearestColor = FALSE,
  ...
)

# S3 method for l_plot3D
loon2ggplot(
  target,
  asAes = TRUE,
  selectedOnTop = TRUE,
  showNearestColor = FALSE,
  ...
)

Arguments

target
a loon or a vector that specifies the widget, layer, glyph, navigator or context completely. The widget is specified by the widget path name (e.g. '.l0.plot'), the remaining objects by their ids.

asAes
logical; set aesthetics attributes, i.e. `color`, `fill` as variables (default TRUE) or general visual properties (FALSE). See details

selectedOnTop
logical and default is TRUE; whether to display the selected points on top. See details.

showNearestColor
logical and default is FALSE; if TRUE, the legend of color and fill (hex code) would be converted to the R built-in color names. For some hex codes, there are no precise matching. Consequently, these colors will be converted to the R built-in color names which are the "nearest" of these hex codes.

...
arguments used inside loon2ggplot(), not used by this method

Value

a ggplot object

Details

in ggplot, generally speaking, there are two ways to set the aesthetics attributes, either take it as variables asAes = TRUE (set in function aes()) or visual properties asAes = FALSE. The main benefits to consider it as variables are that 1. legend could be shown; 2. convenient for further analysis.

In loon, if the selected state is changed, the order of the points will be changed so that the highlighted points will be displayed on top. To turn it static, if selectedOnTop = TRUE, the points will be partitioned into two groups, one group presents the un-selected (un-highlighted) points and the other group presents selected (highlighted) points. the un-selected group will be drawn first, then selected group will be displayed on top of it; if selectedOnTop = FALSE, no partition is applied so that the displayed order will be identical to the original data set order. This is very helpful when further analysis will be operated in ggplot graphical system, e.g. + facet_wrap(...). See examples.

Examples

if(requireNamespace("zenplots") &&
     packageVersion("zenplots") > "1.0.4" &&
     requireNamespace("patchwork")) {
zen <- zenplot(iris, plot1d = "density", pkg = "loon")
ggzen <- loon.ggplot(zen)
ggzen +
  patchwork::plot_annotation(title = "This is a ggplot")
}

#> Loading required namespace: zenplots

if(interactive()) {
lp <- l_plot(iris,
             color = iris$Species,
             glyph = "circle")
gp <- loon2ggplot(lp)
gp # a ggplot object

# add smooth layer, grouped by color
gp +
  geom_smooth(aes(color = color)) +
  # give meaningful legend label names
  scale_color_manual(
    # make sure the order is correct
    values = unique(hex12tohex6(lp['color'])),
    labels = c("setosa", "versicolor", "virginica")
  )

# histogram
lh <- l_hist(mtcars$mpg,
             color = factor(mtcars$gear))

gh0 <- loon2ggplot(lh)
# facet by `fill`
gh0 + facet_wrap(~fill)

# set `asAes = FALSE`
gh1 <- loon2ggplot(lh, asAes = FALSE)
# Expect the legend, they both are identical
gh1
if (FALSE) {
# ERROR
# The bins are constructed by `ggplot2::geom_rect()`
# Very limited operations can be made
gh1 + facet_wrap(~fill)
}

# Argument `selectedOnTop`
p <- l_plot(iris, color = iris$Species)
p['selected'][iris$Petal.Length > 5] <- TRUE
g <- loon.ggplot(p)
# It looks correct.
g
# facet by "Species"
if (FALSE) {
g + facet_wrap(iris$Species)}
# Something is wrong here. There is a pink point (at least one)
# in species "versicolor"! The reason is because after points are
# hightligthed, the displayed order has been changed. One way to
# fix it is to set the `selectedOnTop` as FALSE.

loon.ggplot(p, selectedOnTop = FALSE) +
  facet_wrap(iris$Species)
}